+++ src/main.rs
@@ -14,6 +14,7 @@ extern crate yaml_rust;
 extern crate nom;
 
 use std::env;
+use std::io::Write;
 use std::sync::Arc;
 
 use linefeed::{Interface, ReadResult};
@@ -45,9 +46,14 @@ fn main() {
     let mut sh = shell::Shell::new();
     rcfile::load_rc_files(&mut sh);
 
+    let args: Vec<String> = env::args().collect();
     // this section handles `cicada -c 'echo hi && echo yoo'`,
     // e.g. it could be triggered from Vim (`:!ls` etc).
-    if env::args().len() > 1 {
+    if args.len() > 1 {
+        if args[1] != "-c" {
+            println_stderr!("cicada: run script: to be implemented");
+            return;
+        }
         let line = tools::env_args_to_command_line();
         log!("run with -c args: {}", &line);
         execute::run_procs(&mut sh, &line, false);